+++ src/Components/LoadSpinner/LoadSpinner.tsx
@@ -1,0 +1,10 @@
+export const LoadSpinner = () => {
+  return (
+    <img
+      src="../assets/load-spinner.svg"
+      width="50"
+      height="50"
+      alt="spinner"
+    />
+  );
+};
