+++ src/Components/context/context.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const Context = React.createContext({} as { (url?: string): void });
