+++ src/pages/AppContainer.tsx
@@ -1,15 +1,14 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 import Search from '../Components/Search/Search';
 import '../styles/App.css';
-import ErrorBoundary from '../Components/ErrorBoundary/ErrorBoundary';
 import ErrorButton from '../Components/ErrorButton/ErrorButton';
 import DataviewList from '../Components/DataviewList/DataviewList';
 import { ApiResponse, Person } from '../Components/DataviewItem/types';
 import Pagination from '../Components/Pagination/Pagination';
 import { Routes, Route } from 'react-router-dom';
 import ItemDetails from '../Components/ItemDetails/ItemDetails';
-
-export const Context = React.createContext({} as { (url?: string): void });
+import { Context } from '../Components/context/context';
+import { LoadSpinner } from '../Components/LoadSpinner/LoadSpinner';
 
 function AppContainer() {
   const [responseData, setResponseData] = useState<ApiResponse>();
@@ -60,71 +59,60 @@ function AppContainer() {
       });
   }
 
-  const loadSpinner = (
-    <img
-      src="../assets/load-spinner.svg"
-      width="50"
-      height="50"
-      alt="spinner"
-    />
-  );
-
   return (
-    <ErrorBoundary>
-      <div className="App">
-        <Routes>
-          <Route
-            path="/"
-            element={
-              <div className="flex-container">
-                <div className="content-container">
-                  <Search
-                    updateData={sendSearchQuery}
-                    updateLoadingStatus={updateLoadingStatus}
-                    disabled={isLoading}
-                    value={localStorage.getItem('searchValue') || ''}
-                  />
-                  <ErrorButton />
-                  <hr />
-                  {isLoading ? (
-                    <div className="loadSpinner">{loadSpinner}</div>
-                  ) : (
-                    <div className="dataview-container">
-                      {data?.length ? (
-                        <Context.Provider value={sendDetailsQuery}>
-                          <DataviewList data={data} />
-                        </Context.Provider>
-                      ) : (
-                        <h2 className="">
-                          There is no results for this search
-                        </h2>
-                      )}
-                    </div>
-                  )}
-                  <Pagination
-                    pageCount={pageCount}
-                    changePage={changePage}
-                    previousPage={responseData?.previous}
-                    nextPage={responseData?.next}
-                    hidden={isLoading || !Boolean(data?.length)}
-                  />
-                </div>
-              </div>
-            }
-          >
-            <Route
-              path="/details"
-              element={
-                <ItemDetails
-                  data={detailsData}
-                  isDetailsLoading={isDetailsLoading}
+    <div className="App">
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <div className="flex-container">
+              <div className="content-container">
+                <Search
+                  updateData={sendSearchQuery}
+                  updateLoadingStatus={updateLoadingStatus}
+                  disabled={isLoading}
+                  value={localStorage.getItem('searchValue') || ''}
                 />
-              }
-            />
-          </Route>
-        </Routes>
-      </div>
-    </ErrorBoundary>
+                <ErrorButton />
+                <hr />
+                {isLoading ? (
+                  <div className="loadSpinner">
+                    <LoadSpinner />
+                  </div>
+                ) : (
+                  <div className="dataview-container">
+                    {data?.length ? (
+                      <Context.Provider value={sendDetailsQuery}>
+                        <DataviewList data={data} />
+                      </Context.Provider>
+                    ) : (
+                      <h2 className="">There is no results for this search</h2>
+                    )}
+                  </div>
+                )}
+                <Pagination
+                  pageCount={pageCount}
+                  changePage={changePage}
+                  previousPage={responseData?.previous}
+                  nextPage={responseData?.next}
+                  hidden={isLoading || !Boolean(data?.length)}
+                />
+              </div>
+            </div>
+          }
+        >
+          <Route
+            path="/details"
+            element={
+              <ItemDetails
+                data={detailsData}
+                isDetailsLoading={isDetailsLoading}
+              />
+            }
+          />
+        </Route>
+      </Routes>
+    </div>
   );
 }
